pyproject.toml
[project]
name = "cram-books-mcp"
version = "0.1.0"
description = "MCP server for CRAM Books - connects Claude to Google Sheets"
readme = "README.md"
authors = [
    { name = "ARUOHTA", email = "otaru1214@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.13.1",
    # Google Sheets API direct access
    "google-auth>=2.27.0",
    "google-api-python-client>=2.100.0",
    "gspread>=6.0.0",
    # Environment variable management
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-httpx>=0.35.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "conftest.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]
